Formal changes, bugfixes
