+++ app/Auto/Auto.py
@@ -1,3 +1,4 @@
 import typing as tp
+import close
 
 Sentence = tp.NewType("Sentence", list[str])
