+++ app/FormalSystem/FormalSystem.py
@@ -1,5 +1,6 @@
 from collections import namedtuple
 import typing as tp
+import close
 
 Sentence = tp.NewType("Sentence", list[str])
