+++ app/Lexicon/Lexicon.py
@@ -1,4 +1,5 @@
 import typing as tp
+import close
 
 lexicon = None
