+++ app/Output/Output.py
@@ -1,5 +1,6 @@
 import typing as tp
 from collections import namedtuple
+import close
 
 Sentence = tp.NewType("Sentence", list[str])
 PrintedTree = namedtuple('PrintedTree', ('sentences', 'children', 'closer'))
